requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
pandas==0.24.0
Pillow==5.4.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
torch==1.0.0
torchvision==0.2.1
urllib3==1.24.1
Werkzeug==0.14.1
